+ implemented simple lexer support + implemented Deep Black style from Notepad++ * some refactor - removed unused timer and changed rendering to use floats(better suits 3D, probably changes are incomplete) ! still there is troubles with tabs ! bitmaps should be implemented in order to render indentation guides

+++ src/PositionCache.cxx
@@ -85,7 +85,7 @@ void LineLayout::Resize(int maxLineLength_) {
 		indicators = new char[maxLineLength_ + 1];
 		// Extra position allocated as sometimes the Windows
 		// GetTextExtentExPoint API writes an extra element.
-		positions = new int[maxLineLength_ + 1 + 1];
+		positions = new float[maxLineLength_ + 1 + 1];
 		maxLineLength = maxLineLength_;
 	}
 }
@@ -532,7 +532,7 @@ PositionCacheEntry::PositionCacheEntry() :
 }
 
 void PositionCacheEntry::Set(unsigned int styleNumber_, const char *s_,
-	unsigned int len_, int *positions_, unsigned int clock_) {
+	unsigned int len_, float *positions_, unsigned int clock_) {
 	Clear();
 	styleNumber = styleNumber_;
 	len = len_;
@@ -559,7 +559,7 @@ void PositionCacheEntry::Clear() {
 }
 
 bool PositionCacheEntry::Retrieve(unsigned int styleNumber_, const char *s_,
-	unsigned int len_, int *positions_) const {
+	unsigned int len_, float *positions_) const {
 	if ((styleNumber == styleNumber_) && (len == len_) &&
 		(memcmp(reinterpret_cast<char *>(positions + len), s_, len)== 0)) {
 		for (unsigned int i=0; i<len; i++) {
@@ -624,7 +624,7 @@ void PositionCache::SetSize(size_t size_) {
 }
 
 void PositionCache::MeasureWidths(Surface *surface, ViewStyle &vstyle, unsigned int styleNumber,
-	const char *s, unsigned int len, int *positions) {
+	const char *s, unsigned int len, float *positions) {
 	allClear = false;
 	int probe = -1;
 	if ((size > 0) && (len < 30)) {
